requirements.txt
python-dotenv==1.0.0
tgtg==0.18.2
schedule==1.2.1
